+++ Node.cpp
@@ -8,7 +8,7 @@ Node::Node(int data, std::string name) {
   this->data = data;
   this->name = name;
   //Dynamically allocate neighborMap
-  std::unordered_map<std::string, std::multiset<int>>* mapPointer = new std::unordered_map<std::string, std::multiset<int>>();
+  std::unordered_map<std::string, std::multiset<double>>* mapPointer = new std::unordered_map<std::string, std::multiset<double>>();
   neighborMap = mapPointer;
 }
 
@@ -16,11 +16,11 @@ Node::~Node() {
   delete neighborMap;
 }
 
-void Node::addNeighbor(std::string neighborName, int weight) {
+void Node::addNeighbor(std::string neighborName, double weight) {
   //If the new neighbor is not already a neighbor add it to the list
   if (neighborMap->find(neighborName) == neighborMap->end()) {
-    std::multiset<int> tempSet;
-    std::pair<std::string, std::multiset<int>> tempPair(neighborName,tempSet);
+    std::multiset<double> tempSet;
+    std::pair<std::string, std::multiset<double>> tempPair(neighborName,tempSet);
     neighborMap->insert(tempPair);
   }
 
@@ -32,7 +32,7 @@ int Node::getData() {
   return data;
 }
 
-std::unordered_map<std::string, std::multiset<int>>* Node::getMapPtr() {
+std::unordered_map<std::string, std::multiset<double>>* Node::getMapPtr() {
   return neighborMap;
 }
